fix: Exported functions (e.g. which) may break experiments (FOUNDENG-145) (#351)
Bash-exported functions are set as environment variables and by default
are inherited into singularity containers.   On some systems the which
command is configured this way and injects arguments into the which
command.  When invoked inside of a determined environment image the
which command does not support these arguments and it breaks the check
for the python3 being on the path, thus breaking most experiments.

Clear all exported functions to avoid this potential collision.