chore: Add slurmcluster.sh support for aizn cluster (#648)
The aizn/znode cluster is behind the hpcgate gateway host.
Add support to configure the tunnel properly indirectly through
hpcgate, and pull the auth token indirectly via hpcgate.

OPT_LAUNCHERHOST is the value used in the master.yaml and was always configred
as localhost, so do that automatically instead of requiring it per-node.
Add OPT_LAUNCHERACTUALHOST which is the eventual hostname on which the launcher
is executing to enable indirection via hpcgate.