chore: Increase the timeout for e2e_slurm_restart tests (FOUNDENG-470) (#714)
The _test_master_restart_cmd tests are failing when run as part of the e2e_slurm_restart job inside the nightly workflow. The reason they are failing is that these tests pull an image as part of the test which takes more than a couple of minutes. This delay makes the experiment stay in PULLING state instead of RUNNING state and makes the check for RUNNING status inside _test_master_restart_cmd fail. I increased the timeout from 30 seconds to 300 seconds and tested the e2e_slurm_restart job manually to ensure all the tests run successfully.