fix: Cleanup dispatches associated with terminated allocations on startup [DET-9142] (#750)
On startup active allocations are restored, but for those that are already terminated (for whatever reason) we should cleanup any associated dispatches.   Previously, this code was only used on debug, but if we ever end up in the case where the DispatcherRM failed to handle the allocation termination (due to a bug for example), this will be sure the allocations move from canceling->canceled.