fix: Show more lines when containers haven't started [DET-9285] (#797)
A 2nd part of the problem is when a download is slow and eventually fails
why has it failed, or were there any errors that caused it.
This increment will show up to 500 lines of error log output, but only
if the containers have not all started.   Once the containers start, then
we retain the current limits of most recent 15 lines.