chore: introduce useLoadable hook
this adds a useLoadable hook which is used for reactive requests that

1) aren't polled
2) aren't changed clientside

Doing this manually can be difficult because of edge cases and stylistic
concerns. Consider the following code:

```tsx
const [req, setReq] = useState<Loadable<ReqType>>(NotLoaded)

//...
useEffect(async () => {
  try {
    const result = await makeReq(arg);
    setReq(Loaded(result))
  } catch {
    handleError(e);
  }
}, [arg])
```

this code won't work because effect callbacks are synchronous, so we'll
need to either move the async function to a callback:

```tsx
const [req, setReq] = useState<Loadable<ReqType>>(NotLoaded)
const callMakeReq = useCallback(async () => {
  try {
    const result = await makeReq(arg);
    setReq(Loaded(result))
  } catch {
    handleError(e);
  }
}, [arg])

//...
useEffect(() => {
  callMakeReq();
}, [callMakeReq])
```

which creates an indirect dependency chain (that is, the effect is
really dependent on the dependencies of the callback, not the callback
itself), or use an iife:

```tsx
const [req, setReq] = useState<Loadable<ReqType>>(NotLoaded)

//...
useEffect(() => {
    (async () => {
      try {
        const result = await makeReq(arg);
        setReq(Loaded(result))
      } catch {
        handleError(e);
      }
    })()
}, [arg])
```

which feels kind of odd and reduces the horizontal space left before we
format the code.

In addition, all of the code examples above have a race condition.
Nothing except for timing guarantees that calls of `makeReq(arg)` return
in order, so a longer call can wipe out the state change from a shorter
call if the shorter call comes after the longer call, so we'd need to
mark the closure as stale or unmounted by using the `useEffect` callback
return value:

```tsx
const [req, setReq] = useState<Loadable<ReqType>>(NotLoaded)

//...
useEffect(() => {
  let mounted = true
  (async () => {
    try {
      const result = await makeReq(arg);
      if (mounted) {
        setReq(Loaded(result))
      }
    } catch {
      handleError(e);
    }
  })()
  return () => { mounted = false }
}, [arg])
```

`useLoadable` fixes both these problems. It takes an async function and
handles the race condition while also wrapping the value in `Loadable`
if necessary. As a nice bonus, the function is typed such that it should
be able to read the function return type:

```tsx
const req = useLoadable(async () => {
  try {
    return await makeReq(arg);
  } catch {
    handleError(e);
    return NotLoaded;
  }
})
```

A small quirk is that the error branch needs to return `NotLoaded` so
the hook doesn't assign the return value as a Loadable<undefined>. In
addition, the hook uses an AbortController to handle whether or not to
set the state. This is passed in as an argument to the callback, meaning
if your async function uses an api call, it can be cancelled at that
point using it:
```tsx
const req = useLoadable(async (canceler) => {
  try {
    return await makeReq(arg, { signal: canceler.signal });
  } catch {
    handleError(e);
    return NotLoaded;
  }
})
```