feat: new user management filters
This updates the user management table to have more and clearer filter
functionality. The name filter is extracted out from the header menu
into a search bar above the table, and role and status filters are added
as well. There's a refactor of the user management table as well --
we've removed the dependency on `useSettings`, instead reading/writing
directly from tue user store. In addition, we no longer read the user
data from the user store, instead doing filtering and sorting on the
backend.

As there are other places where we have this pattern (read-only
remote data is requested only when dependent values change), we
introduce a new `useLoadable` hook which handles holding the internal
state, throwing out stale updates, and setting up cancelers while also
allowing us to use the `async/await` style we're used to. The only
oddity with the API is that the hook infers its return type from the
function passed in -- when handling errors or otherwise indicating that
the result shouldn't be loaded, the user has to return `NotLoaded`
instead of not returning anything.